+++ R/polished_config.R
@@ -36,6 +36,8 @@
 #' Set to \code{NULL} to force Sign Out at session end. This argument is passed to
 #' the `expires` option in js-cookie: \url{https://github.com/js-cookie/js-cookie#expires}.
 #' Default value is \code{365L} (i.e. 1 year)
+#' @param is_two_fa_required boolean specifying whether or not 2 factor authentication is required.  Defaults
+#' to \code{FALSE}.
 #'
 #' @export
 #'
@@ -72,7 +74,8 @@ polished_config <- function(
   is_email_verification_required = TRUE,
   sentry_dsn = NULL,
   cookie_expires = 365L,
-  is_auth_required = TRUE
+  is_auth_required = TRUE,
+  is_two_fa_required = FALSE
 ) {
 
   if (!(length(api_key) == 1 && is.character(api_key))) {
@@ -126,6 +129,9 @@ call. = FALSE
   if (!(length(is_email_verification_required) == 1 && is.logical(is_email_verification_required))) {
     stop("invalid `is_email_verification_required` argument passed to `polished_config()`", call. = FALSE)
   }
+  if (!(length(is_two_fa_required) == 1 && is.logical(is_two_fa_required))) {
+    stop("invalid `is_two_fa_required` argument passed to `polished_config()`", call. = FALSE)
+  }
 
 
 
@@ -140,6 +146,7 @@ call. = FALSE
   assign("sentry_dsn", sentry_dsn, envir = .polished)
   assign("cookie_expires", cookie_expires, envir = .polished)
   assign("is_auth_required", is_auth_required, envir = .polished)
+  assign("is_two_fa_required", is_two_fa_required, envir = .polished)
 
 
   if (!is.null(firebase_config)) {