+++ inst/examples/polished_example_03/global.R
@@ -6,7 +6,7 @@ library(shinyjs)
 app_config <- config::get()
 
 # configure polished
-Polished$new(
+polished_config(
   app_name = "polished_example_03",
   api_key = app_config$api_key,
   is_auth_required = FALSE